+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Changed]
 ### [Deprecated]
 ### [Removed]
+- @aofl/router - removed forceReload option and can't navigate to current route restriction
 ### [Fixed]
 - @aofl/cli-lib/git - archive paths reference
 ### [Security]