+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased](https://github.com/AgeOfLearning/aofl/compare/v3.8.1...development) - Date
+## [unreleased](https://github.com/AgeOfLearning/aofl/compare/v3.9.0...development) - Date
+
+### [Added]
+### [Changed]
+### [Deprecated]
+### [Removed]
+### [Fixed]
+### [Security]
+
+---
+## [v3.9.0](https://github.com/AgeOfLearning/aofl/compare/v3.8.1...v3.9.0) - 2020-05-15
 
 ### [Added]
 - @aofl/cli - init will invoke `postinit` script